Add database/firebaseConfig.js
